Add files via upload
Final version!
This is the newest main file. A cable is only needed to transmit the code to the Micro:bit. After, the cable can be taken away and the Micro:bit gets attached to the wrist. During the night the Micro:bit checks the seconds until the time you need to get up. The next day you must flash the code onto the Micro:bit once more.